requirements.txt
# --- Requisitos principales ---
fastapi
uvicorn

# --- Modelos de visión (YOLO y PIL) ---
ultralytics
pillow

# --- IA y manejo de documentos ---
openai
python-dotenv
pymupdf  # (fitz) para extraer texto de PDF
base64io  # para manejo base64 si deseas manipular flujos binarios (opcional)

# --- Tipado MIME ---
python-magic  # útil para detectar tipo MIME con más precisión (opcional)
